+++ src/games/loa.ts
@@ -39,22 +39,20 @@ export class LinesOfActionGame extends GameBase {
             {
                 uid: "scrambled",
                 group: "setup"
+            },
+            {
+                uid: "classic",
+                group: "board"
             }
         ],
         flags: ["multistep", "check"]
     };
 
-    public static coords2algebraic(x: number, y: number): string {
-        return GameBase.coords2algebraic(x, y, 9);
-    }
-    public static algebraic2coords(cell: string): [number, number] {
-        return GameBase.algebraic2coords(cell, 9);
-    }
-
     public numplayers = 2;
     public currplayer: playerid = 1;
     public board!: Map<string, playerid>;
     public gameover = false;
+    public boardsize = 9;
     public winner: playerid[] = [];
     public variants: string[] = [];
     public stack!: Array<IMoveState>;
@@ -64,21 +62,45 @@ export class LinesOfActionGame extends GameBase {
     constructor(state?: ILinesOfActionState | string, variants?: string[]) {
         super();
         if (state === undefined) {
-            let board = new Map<string, playerid>([
-                ["b9", 1], ["c9", 1], ["d9", 1], ["e9", 1], ["f9", 1], ["g9", 1], ["h9", 1],
-                ["b1", 1], ["c1", 1], ["d1", 1], ["e1", 1], ["f1", 1], ["g1", 1], ["h1", 1],
-                ["a2", 2], ["a3", 2], ["a4", 2], ["a5", 2], ["a6", 2], ["a7", 2], ["a8", 2],
-                ["i2", 2], ["i3", 2], ["i4", 2], ["i5", 2], ["i6", 2], ["i7", 2], ["i8", 2],
-            ]);
-            if ( (variants !== undefined) && (variants.length === 1) && (variants[0] === "scrambled") ) {
-                this.variants = ["scrambled"];
-                board = new Map<string, playerid>([
-                    ["b9", 1], ["c9", 2], ["d9", 1], ["e9", 2], ["f9", 1], ["g9", 2], ["h9", 1],
-                    ["b1", 1], ["c1", 2], ["d1", 1], ["e1", 2], ["f1", 1], ["g1", 2], ["h1", 1],
-                    ["a2", 2], ["a3", 1], ["a4", 2], ["a5", 1], ["a6", 2], ["a7", 1], ["a8", 2],
-                    ["i2", 2], ["i3", 1], ["i4", 2], ["i5", 1], ["i6", 2], ["i7", 1], ["i8", 2],
-                ]);
+            if ( (variants !== undefined) && (variants.length > 0) ) {
+                this.variants = [...variants];
             }
+            let board: Map<string,playerid>;
+            if (this.variants.includes("classic")) {
+                this.boardsize = 8;
+                if (this.variants.includes("scrambled")) {
+                    board = new Map<string, playerid>([
+                        ["b8", 1], ["c8", 2], ["d8", 1], ["e8", 2], ["f8", 1], ["g8", 2],
+                        ["b1", 2], ["c1", 1], ["d1", 2], ["e1", 1], ["f1", 2], ["g1", 1],
+                        ["a2", 1], ["a3", 2], ["a4", 1], ["a5", 2], ["a6", 1], ["a7", 2],
+                        ["h2", 2], ["h3", 1], ["h4", 2], ["h5", 1], ["h6", 2], ["h7", 1],
+                    ]);
+                } else {
+                    board = new Map<string, playerid>([
+                        ["b8", 1], ["c8", 1], ["d8", 1], ["e8", 1], ["f8", 1], ["g8", 1],
+                        ["b1", 1], ["c1", 1], ["d1", 1], ["e1", 1], ["f1", 1], ["g1", 1],
+                        ["a2", 2], ["a3", 2], ["a4", 2], ["a5", 2], ["a6", 2], ["a7", 2],
+                        ["h2", 2], ["h3", 2], ["h4", 2], ["h5", 2], ["h6", 2], ["h7", 2],
+                    ]);
+                }
+            } else {
+                if (this.variants.includes("scrambled")) {
+                    board = new Map<string, playerid>([
+                        ["b9", 1], ["c9", 2], ["d9", 1], ["e9", 2], ["f9", 1], ["g9", 2], ["h9", 1],
+                        ["b1", 1], ["c1", 2], ["d1", 1], ["e1", 2], ["f1", 1], ["g1", 2], ["h1", 1],
+                        ["a2", 2], ["a3", 1], ["a4", 2], ["a5", 1], ["a6", 2], ["a7", 1], ["a8", 2],
+                        ["i2", 2], ["i3", 1], ["i4", 2], ["i5", 1], ["i6", 2], ["i7", 1], ["i8", 2],
+                    ]);
+                } else {
+                    board = new Map<string, playerid>([
+                        ["b9", 1], ["c9", 1], ["d9", 1], ["e9", 1], ["f9", 1], ["g9", 1], ["h9", 1],
+                        ["b1", 1], ["c1", 1], ["d1", 1], ["e1", 1], ["f1", 1], ["g1", 1], ["h1", 1],
+                        ["a2", 2], ["a3", 2], ["a4", 2], ["a5", 2], ["a6", 2], ["a7", 2], ["a8", 2],
+                        ["i2", 2], ["i3", 2], ["i4", 2], ["i5", 2], ["i6", 2], ["i7", 2], ["i8", 2],
+                    ]);
+                }
+            }
+
             const fresh: IMoveState = {
                 _version: LinesOfActionGame.gameinfo.version,
                 _results: [],
@@ -115,6 +137,12 @@ export class LinesOfActionGame extends GameBase {
         this.board = new Map(state.board);
         this.lastmove = state.lastmove;
         this.results = [...state._results];
+
+        if (this.variants.includes("classic")) {
+            this.boardsize = 8;
+        } else {
+            this.boardsize = 9;
+        }
         return this;
     }
 
@@ -125,18 +153,18 @@ export class LinesOfActionGame extends GameBase {
         }
         const moves: string[] = [];
 
-        const grid = new RectGrid(9, 9);
+        const grid = new RectGrid(this.boardsize, this.boardsize);
         const dirPairs: [Directions, Directions][] = [["N", "S"], ["E", "W"], ["NE", "SW"], ["NW", "SE"]];
         const pieces = [...this.board.entries()].filter(e => e[1] === player).map(e => e[0]);
         for (const cell of pieces) {
-            const [x, y] = LinesOfActionGame.algebraic2coords(cell);
+            const [x, y] = LinesOfActionGame.algebraic2coords(cell, this.boardsize);
             for (const pair of dirPairs) {
                 const rays: [number, number][][] = [];
                 let magnitude = 1;
                 for (const d of pair) {
                     const ray = grid.ray(x, y, d);
                     for (const point of ray) {
-                        if (this.board.has(LinesOfActionGame.coords2algebraic(...point))) {
+                        if (this.board.has(LinesOfActionGame.coords2algebraic(...point, this.boardsize))) {
                             magnitude++;
                         }
                     }
@@ -147,7 +175,7 @@ export class LinesOfActionGame extends GameBase {
                         let valid = true;
                         for (let i = 0; i < magnitude - 1; i++) {
                             const next = ray[i];
-                            const nextCell = LinesOfActionGame.coords2algebraic(...next);
+                            const nextCell = LinesOfActionGame.coords2algebraic(...next, this.boardsize);
                             if (this.board.has(nextCell)) {
                                 const contents = this.board.get(nextCell);
                                 if (contents !== player) {
@@ -158,7 +186,7 @@ export class LinesOfActionGame extends GameBase {
                         }
                         if (valid) {
                             const next = ray[magnitude - 1];
-                            const nextCell = LinesOfActionGame.coords2algebraic(...next);
+                            const nextCell = LinesOfActionGame.coords2algebraic(...next, this.boardsize);
                             if (this.board.has(nextCell)) {
                                 const contents = this.board.get(nextCell);
                                 if (contents !== player) {
@@ -186,11 +214,11 @@ export class LinesOfActionGame extends GameBase {
             return undefined;
         }
         const targets: string[] = [];
-        const grid = new RectGrid(9, 9);
-        const [x, y] = LinesOfActionGame.algebraic2coords(start);
+        const grid = new RectGrid(this.boardsize, this.boardsize);
+        const [x, y] = LinesOfActionGame.algebraic2coords(start, this.boardsize);
         for (const pair of [["N", "S"] as const, ["E", "W"] as const, ["NE", "SW"] as const, ["NW", "SE"] as const]) {
-            const ray1 = grid.ray(x, y, pair[0]).map(pt => LinesOfActionGame.coords2algebraic(...pt));
-            const ray2 = grid.ray(x, y, pair[1]).map(pt => LinesOfActionGame.coords2algebraic(...pt));
+            const ray1 = grid.ray(x, y, pair[0]).map(pt => LinesOfActionGame.coords2algebraic(...pt, this.boardsize));
+            const ray2 = grid.ray(x, y, pair[1]).map(pt => LinesOfActionGame.coords2algebraic(...pt, this.boardsize));
             const combined: string[] = [start, ...ray1, ...ray2];
             const numPieces = [...this.board.entries()].filter(e => combined.includes(e[0])).length;
             for (const ray of [ray1, ray2]) {
@@ -217,7 +245,7 @@ export class LinesOfActionGame extends GameBase {
 
     public handleClick(move: string, row: number, col: number, piece?: string): IClickResult {
         try {
-            const cell = LinesOfActionGame.coords2algebraic(col, row);
+            const cell = LinesOfActionGame.coords2algebraic(col, row, this.boardsize);
             let newmove = "";
             if (move.length > 0) {
                 let prev = move;
@@ -276,7 +304,7 @@ export class LinesOfActionGame extends GameBase {
         if (from !== undefined) {
             // cell is valid
             try {
-                LinesOfActionGame.algebraic2coords(from);
+                LinesOfActionGame.algebraic2coords(from, this.boardsize);
             } catch {
                 result.valid = false;
                 result.message = i18next.t("apgames:validation._general.INVALIDCELL", {cell: from});
@@ -310,7 +338,7 @@ export class LinesOfActionGame extends GameBase {
             } else {
                 // cell is valid
                 try {
-                    LinesOfActionGame.algebraic2coords(to);
+                    LinesOfActionGame.algebraic2coords(to, this.boardsize);
                 } catch {
                     result.valid = false;
                     result.message = i18next.t("apgames:validation._general.INVALIDCELL", {cell: to});
@@ -362,7 +390,7 @@ export class LinesOfActionGame extends GameBase {
             const [cell,] = m.split(/[-x]/);
             const pts = this.findPoints(cell);
             if (pts !== undefined) {
-                this._points = pts.map(c => LinesOfActionGame.algebraic2coords(c));
+                this._points = pts.map(c => LinesOfActionGame.algebraic2coords(c, this.boardsize));
             } else {
                 this._points = [];
             }
@@ -375,11 +403,11 @@ export class LinesOfActionGame extends GameBase {
         this.results = [];
         const [from, to] = m.split(/[-x]/);
         this.board.delete(from);
-        if (to !== "e5") {
+        if ( (this.variants.includes("classic")) || (to !== "e5") ) {
             this.board.set(to, this.currplayer);
         }
         this.results.push({type: "move", from, to});
-        if (to === "e5") {
+        if ( (! this.variants.includes("classic")) && (to === "e5") ) {
             this.results.push({type: "capture", where: "e5"})
         } else if (m.includes("x")) {
             this.results.push({type: "capture", where: to})
@@ -424,16 +452,16 @@ export class LinesOfActionGame extends GameBase {
 
     private isConnected(player: playerid): boolean {
         const pieces = [...this.board.entries()].filter(e => e[1] === player).map(e => e[0]);
-        const grid = new RectGrid(9, 9);
+        const grid = new RectGrid(this.boardsize, this.boardsize);
         const seen: Set<string> = new Set();
         const todo: string[] = [pieces[0]];
         while (todo.length > 0) {
             const cell = todo.pop();
             seen.add(cell!);
-            const [x, y] = LinesOfActionGame.algebraic2coords(cell!);
+            const [x, y] = LinesOfActionGame.algebraic2coords(cell!, this.boardsize);
             const neighbours = grid.adjacencies(x, y);
             for (const n of neighbours) {
-                const nCell = LinesOfActionGame.coords2algebraic(...n);
+                const nCell = LinesOfActionGame.coords2algebraic(...n, this.boardsize);
                 if (pieces.includes(nCell)) {
                     if (! seen.has(nCell)) {
                         todo.push(nCell);
@@ -448,7 +476,7 @@ export class LinesOfActionGame extends GameBase {
         return {
             game: LinesOfActionGame.gameinfo.uid,
             numplayers: this.numplayers,
-            variants: this.variants,
+            variants: [...this.variants],
             gameover: this.gameover,
             winner: [...this.winner],
             stack: [...this.stack]
@@ -469,13 +497,13 @@ export class LinesOfActionGame extends GameBase {
     public render(): APRenderRep {
         // Build piece string
         let pstr = "";
-        for (let row = 0; row < 9; row++) {
+        for (let row = 0; row < this.boardsize; row++) {
             if (pstr.length > 0) {
                 pstr += "\n";
             }
             const pieces: string[] = [];
-            for (let col = 0; col < 9; col++) {
-                const cell = LinesOfActionGame.coords2algebraic(col, row);
+            for (let col = 0; col < this.boardsize; col++) {
+                const cell = LinesOfActionGame.coords2algebraic(col, row, this.boardsize);
                 if (this.board.has(cell)) {
                     const contents = this.board.get(cell);
                     if (contents === undefined) {
@@ -486,7 +514,7 @@ export class LinesOfActionGame extends GameBase {
                     } else {
                         pieces.push("B");
                     }
-                } else if (cell === "e5") {
+                } else if ( (! this.variants.includes("classic")) && (cell === "e5") ) {
                     pieces.push("X");
                 } else {
                     pieces.push("-");
@@ -494,14 +522,18 @@ export class LinesOfActionGame extends GameBase {
             }
             pstr += pieces.join("");
         }
-        pstr = pstr.replace(/-{9}/g, "_");
+        if (this.variants.includes("classic")) {
+            pstr = pstr.replace(/-{8}/g, "_");
+        } else {
+            pstr = pstr.replace(/-{9}/g, "_");
+        }
 
         // Build rep
         const rep: APRenderRep =  {
             board: {
                 style: "squares-checkered",
-                width: 9,
-                height: 9,
+                width: this.boardsize,
+                height: this.boardsize,
             },
             legend: {
                 A: {
@@ -519,6 +551,9 @@ export class LinesOfActionGame extends GameBase {
             },
             pieces: pstr
         };
+        if (this.variants.includes("classic")) {
+            delete rep.legend!.X;
+        }
 
         // Add annotations
         if ( (this.results.length > 0) || (this._points.length > 0) ){
@@ -536,11 +571,11 @@ export class LinesOfActionGame extends GameBase {
 
             for (const move of this.results) {
                 if (move.type === "move") {
-                    const [fromX, fromY] = LinesOfActionGame.algebraic2coords(move.from);
-                    const [toX, toY] = LinesOfActionGame.algebraic2coords(move.to);
+                    const [fromX, fromY] = LinesOfActionGame.algebraic2coords(move.from, this.boardsize);
+                    const [toX, toY] = LinesOfActionGame.algebraic2coords(move.to, this.boardsize);
                     rep.annotations.push({type: "move", targets: [{row: fromY, col: fromX}, {row: toY, col: toX}]});
                 } else if (move.type === "capture") {
-                    const [x, y] = LinesOfActionGame.algebraic2coords(move.where!);
+                    const [x, y] = LinesOfActionGame.algebraic2coords(move.where!, this.boardsize);
                     rep.annotations.push({type: "exit", targets: [{row: y, col: x}]});
                 }
             }